Move authorisedRawRequest into authorised-request (no refactor yet)
